Ano3/PECI/projeto/Backend/api/serializers.py
from rest_framework import serializers
from .models import Accommodation, Restaurant

class AccommodationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Accommodation
        fields = '__all__'

class RestaurantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Restaurant
        fields = '__all__'
