Ano3/PECI/projeto/Backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
pymongo==3.11.0
python-dateutil==2.8.2
pytz==2022.7.1
setuptools==59.6.0
six==1.16.0
sqlparse==0.2.4